Promote EffectsManager.ColoredGlyphState out of EManager
