+++ SadConsole/Effects/Blink.cs
@@ -52,7 +52,7 @@ namespace SadConsole.Effects
         }
 
         /// <inheritdoc />
-        public override bool ApplyToCell(ColoredGlyph cell, EffectsManager.ColoredGlyphState originalState)
+        public override bool ApplyToCell(ColoredGlyph cell, ColoredGlyphState originalState)
         {
             Color oldColor = cell.Foreground;
