+++ SadConsole/Effects/Recolor.cs
@@ -45,7 +45,7 @@ namespace SadConsole.Effects
         }
 
         /// <inheritdoc />
-        public override bool ApplyToCell(ColoredGlyph cell, EffectsManager.ColoredGlyphState originalState)
+        public override bool ApplyToCell(ColoredGlyph cell, ColoredGlyphState originalState)
         {
             Color oldForeground = cell.Foreground;
             Color oldBackground = cell.Background;
